+++ datagrowth/configuration/__init__.py
@@ -1,7 +1,7 @@
 from datagrowth.configuration.types import (ConfigurationType, ConfigurationProperty, ConfigurationNotFoundError,
                                             create_config, register_defaults)
 from datagrowth.configuration.fields import ConfigurationField, ConfigurationFormField
-from datagrowth.configuration.serializers import load_config, DecodeConfigAction, get_standardized_configuration
+from datagrowth.configuration.serializers import load_config, DecodeConfigAction
 
 
 DATAGROWTH_CONFIGURATION = ConfigurationType()