+++ tests/django_project/configuration/tests.py
@@ -6,7 +6,7 @@ from django.test import TestCase
 
 from datagrowth.configuration.defaults import DATAGROWTH_DEFAULT_CONFIGURATION
 from datagrowth.configuration import (ConfigurationType, ConfigurationNotFoundError, ConfigurationProperty, load_config,
-                                      get_standardized_configuration, create_config, register_defaults)
+                                      create_config, register_defaults)
 
 
 MOCK_CONFIGURATION = deepcopy(DATAGROWTH_DEFAULT_CONFIGURATION)
@@ -385,24 +385,6 @@ class TestLoadConfigDecorator(TestCase):
             self.fail("load_config did not throw an exception when no config kwarg was set.")
         except TypeError:
             pass
-
-
-class TestGetStandardizedConfiguration(TestCase):
-
-    def setUp(self):
-        super().setUp()
-        self.config = ConfigurationType(namespace="name", private=["_test3"], defaults=MOCK_CONFIGURATION)
-        self.config.update({
-            "test": "public",
-            "_test2": "protected",
-            "_test3": "private"
-        })
-
-    def test_standardized_configuration(self):
-        out = get_standardized_configuration(self.config)
-        self.assertEqual(out, "3601ce2b866f9ccff5e9e49b628e65108abb3d5ada72fce6511645212c0ce520")
-        out = get_standardized_configuration(self.config, as_hash=False)
-        self.assertEqual(out, "test=public")
 
 
 class TestCreateConfig(TestCase):